feat: init.vimからinit.luaへの完全移行
